requirements.txt
Flask
requests
slack_sdk
google-api-python-client
google-auth
google-auth-oauthlib
